meging opus and characters apps to writings app finished
