+++ dev_data.json
@@ -1,1 +1,239 @@
-[{"model": "opus.author", "pk": 1, "fields": {"name": "Лев Толстой", "birth_date": null, "death_date": null}}, {"model": "opus.author", "pk": 2, "fields": {"name": "Достоевский Фёдор Михайлович", "birth_date": null, "death_date": null}}, {"model": "opus.opus", "pk": 1, "fields": {"name": "Война и Мир", "author": 1, "publish_date": null}}, {"model": "opus.opus", "pk": 2, "fields": {"name": "Преступление и наказание", "author": 2, "publish_date": null}}, {"model": "opus.opus", "pk": 3, "fields": {"name": "Идиот", "author": 2, "publish_date": null}}, {"model": "characters.character", "pk": 1, "fields": {"name": "Андрей Болконский", "gender": "M", "opus": 1}}, {"model": "characters.character", "pk": 2, "fields": {"name": "Пьер Безухов", "gender": "M", "opus": 1}}, {"model": "characters.character", "pk": 3, "fields": {"name": "Родион Раскольников", "gender": "M", "opus": 2}}, {"model": "characters.character", "pk": 4, "fields": {"name": "Князь Мышкин", "gender": "M", "opus": 3}}, {"model": "characters.quote", "pk": 1, "fields": {"name": null, "source": 2, "text": "Надо жить, надо любить, надо верить."}}, {"model": "characters.quote", "pk": 2, "fields": {"name": null, "source": 4, "text": "Я не был влюблен... я... был счастлив иначе."}}, {"model": "characters.quote", "pk": 3, "fields": {"name": null, "source": 4, "text": "Знаешь ли, что женщина способна замучить человека жестокостями и насмешками и ни разу угрызения совести не почувствует, потому что про себя каждый раз будет думать, смотря на тебя: «Вот теперь я его измучаю до смерти, да зато потом ему любовью моей наверстаю...»"}}, {"model": "characters.quote", "pk": 4, "fields": {"name": null, "source": 1, "text": "Последнее время мне стало жить тяжело. Я вижу, я стал понимать слишком много."}}, {"model": "characters.quote", "pk": 5, "fields": {"name": null, "source": 1, "text": "Сражение выигрывает тот, кто твердо решил его выиграть!"}}, {"model": "characters.quote", "pk": 6, "fields": {"name": null, "source": 2, "text": "Ничем не может владеть человек, пока боится смерти. А кто не боится её, тому принадлежит всё."}}, {"model": "characters.quote", "pk": 7, "fields": {"name": null, "source": 3, "text": "Человек он умный, но чтоб умно поступать — одного ума мало."}}, {"model": "characters.quote", "pk": 8, "fields": {"name": null, "source": 3, "text": "Смогу ли я переступить или не смогу! Осмелюсь ли нагнуться и взять или нет? Тварь ли я дрожащая или право имею!"}}, {"model": "tests.quotequestion", "pk": 1, "fields": {"quote": 2, "alt_answers": [1, 2, 3]}}, {"model": "tests.quotequestion", "pk": 2, "fields": {"quote": 1, "alt_answers": [1, 3, 4]}}, {"model": "tests.quotequestion", "pk": 3, "fields": {"quote": 3, "alt_answers": [1, 2, 3]}}, {"model": "tests.quotequestion", "pk": 4, "fields": {"quote": 4, "alt_answers": [2, 3, 4]}}, {"model": "tests.quotequestion", "pk": 5, "fields": {"quote": 5, "alt_answers": [2, 3, 4]}}, {"model": "tests.quotequestion", "pk": 6, "fields": {"quote": 6, "alt_answers": [1, 3, 4]}}, {"model": "tests.quotequestion", "pk": 7, "fields": {"quote": 7, "alt_answers": [1, 2, 4]}}]
+[
+{
+  "model": "writings.author",
+  "pk": 1,
+  "fields": {
+    "name": "Лев Толстой",
+    "birth_date": null,
+    "death_date": null
+  }
+},
+{
+  "model": "writings.author",
+  "pk": 2,
+  "fields": {
+    "name": "Достоевский Фёдор Михайлович",
+    "birth_date": null,
+    "death_date": null
+  }
+},
+{
+  "model": "writings.writing",
+  "pk": 1,
+  "fields": {
+    "name": "Война и Мир",
+    "author": 1,
+    "publish_date": null
+  }
+},
+{
+  "model": "writings.writing",
+  "pk": 2,
+  "fields": {
+    "name": "Преступление и наказание",
+    "author": 2,
+    "publish_date": null
+  }
+},
+{
+  "model": "writings.writing",
+  "pk": 3,
+  "fields": {
+    "name": "Идиот",
+    "author": 2,
+    "publish_date": null
+  }
+},
+{
+  "model": "writings.character",
+  "pk": 1,
+  "fields": {
+    "name": "Андрей Болконский",
+    "gender": "M",
+    "writing": 1
+  }
+},
+{
+  "model": "writings.character",
+  "pk": 2,
+  "fields": {
+    "name": "Пьер Безухов",
+    "gender": "M",
+    "writing": 1
+  }
+},
+{
+  "model": "writings.character",
+  "pk": 3,
+  "fields": {
+    "name": "Родион Раскольников",
+    "gender": "M",
+    "writing": 2
+  }
+},
+{
+  "model": "writings.character",
+  "pk": 4,
+  "fields": {
+    "name": "Князь Мышкин",
+    "gender": "M",
+    "writing": 3
+  }
+},
+{
+  "model": "writings.quote",
+  "pk": 1,
+  "fields": {
+    "name": null,
+    "source": 2,
+    "text": "Надо жить, надо любить, надо верить."
+  }
+},
+{
+  "model": "writings.quote",
+  "pk": 2,
+  "fields": {
+    "name": null,
+    "source": 4,
+    "text": "Я не был влюблен... я... был счастлив иначе."
+  }
+},
+{
+  "model": "writings.quote",
+  "pk": 3,
+  "fields": {
+    "name": null,
+    "source": 4,
+    "text": "Знаешь ли, что женщина способна замучить человека жестокостями и насмешками и ни разу угрызения совести не почувствует, потому что про себя каждый раз будет думать, смотря на тебя: «Вот теперь я его измучаю до смерти, да зато потом ему любовью моей наверстаю...»"
+  }
+},
+{
+  "model": "writings.quote",
+  "pk": 4,
+  "fields": {
+    "name": null,
+    "source": 1,
+    "text": "Последнее время мне стало жить тяжело. Я вижу, я стал понимать слишком много."
+  }
+},
+{
+  "model": "writings.quote",
+  "pk": 5,
+  "fields": {
+    "name": null,
+    "source": 1,
+    "text": "Сражение выигрывает тот, кто твердо решил его выиграть!"
+  }
+},
+{
+  "model": "writings.quote",
+  "pk": 6,
+  "fields": {
+    "name": null,
+    "source": 2,
+    "text": "Ничем не может владеть человек, пока боится смерти. А кто не боится её, тому принадлежит всё."
+  }
+},
+{
+  "model": "writings.quote",
+  "pk": 7,
+  "fields": {
+    "name": null,
+    "source": 3,
+    "text": "Человек он умный, но чтоб умно поступать — одного ума мало."
+  }
+},
+{
+  "model": "writings.quote",
+  "pk": 8,
+  "fields": {
+    "name": null,
+    "source": 3,
+    "text": "Смогу ли я переступить или не смогу! Осмелюсь ли нагнуться и взять или нет? Тварь ли я дрожащая или право имею!"
+  }
+},
+{
+  "model": "tests.quotequestion",
+  "pk": 1,
+  "fields": {
+    "quote": 2,
+    "alt_answers": [
+      1,
+      2,
+      3
+    ]
+  }
+},
+{
+  "model": "tests.quotequestion",
+  "pk": 2,
+  "fields": {
+    "quote": 1,
+    "alt_answers": [
+      1,
+      3,
+      4
+    ]
+  }
+},
+{
+  "model": "tests.quotequestion",
+  "pk": 3,
+  "fields": {
+    "quote": 3,
+    "alt_answers": [
+      1,
+      2,
+      3
+    ]
+  }
+},
+{
+  "model": "tests.quotequestion",
+  "pk": 4,
+  "fields": {
+    "quote": 4,
+    "alt_answers": [
+      2,
+      3,
+      4
+    ]
+  }
+},
+{
+  "model": "tests.quotequestion",
+  "pk": 5,
+  "fields": {
+    "quote": 5,
+    "alt_answers": [
+      2,
+      3,
+      4
+    ]
+  }
+},
+{
+  "model": "tests.quotequestion",
+  "pk": 6,
+  "fields": {
+    "quote": 6,
+    "alt_answers": [
+      1,
+      3,
+      4
+    ]
+  }
+},
+{
+  "model": "tests.quotequestion",
+  "pk": 7,
+  "fields": {
+    "quote": 7,
+    "alt_answers": [
+      1,
+      2,
+      4
+    ]
+  }
+}
+]
