--- litest/opus/admin.py
@@ -1,7 +1,0 @@
-from django.contrib import admin
-from .models import Opus, Author
-
-admin.site.register((
-    Opus,
-    Author,
-))
