+++ litest/writings/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class CharactersConfig(AppConfig):
+class WritingsConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'characters'
+    name = 'writings'
