+++ litest/writings/models.py
@@ -1,5 +1,32 @@
 from django.db import models
-from opus.models import Opus
+
+
+class Author(models.Model):
+    """Model of authors of writings"""
+    name = models.CharField(max_length=64, verbose_name="Имя")
+    birth_date = models.DateField(null=True, blank=True, verbose_name="Дата рождения")
+    death_date = models.DateField(null=True, blank=True, verbose_name="Дата смерти")
+
+    class Meta:
+        verbose_name = "Автор"
+        verbose_name_plural = "Авторы"
+
+    def __str__(self) -> str:
+        return self.name
+
+
+class Writing(models.Model):
+    """Model of writings"""
+    name = models.CharField(max_length=128, verbose_name="Название")
+    author = models.ForeignKey(Author, on_delete=models.CASCADE, verbose_name="Автор")
+    publish_date = models.DateField(null=True, blank=True, verbose_name="Дата публикации")
+
+    class Meta:
+        verbose_name = "Произведение"
+        verbose_name_plural = "Произведения"
+
+    def __str__(self) -> str:
+        return self.name
 
 
 class Character(models.Model):
@@ -9,7 +36,7 @@ class Character(models.Model):
     )
     name = models.CharField(max_length=64, verbose_name="Имя")
     gender = models.CharField(choices=GENDERS, max_length=1, verbose_name="Пол")
-    opus = models.ForeignKey(Opus, on_delete=models.CASCADE, verbose_name="Произведение")
+    writing = models.ForeignKey(Writing, on_delete=models.CASCADE, verbose_name="Произведение")
 
     class Meta:
         verbose_name = "Персонаж"
